Fix codegen when using SureFire Booter combined with JaCoCo
When using codegen with SureFire Booter + JaCoCo, codegen was unable to
determine the correct classpath, as its check for SureFire Booter was too
strict (both the surefirebooter and jacoco jars were on the cp, while the
code was only expecting surefirebooter).

Fix this in a more generic way by checking all classloader URLs for
surefirebooter.