Adds methods to SQLListenerContext to retrieve SQLBindings.
This was added as an alternative to just getting an SQL string. Internally,
SQLListenerContextImpl now only takes SQLBindings and the getSQL and
getSQLStatements methods delegate to SQLBindings#getSQL. This required
shuffling around AbstractSQLQuery and SQL*Clause to create bindings sooner to
pass through to addSQL even though we derive the same data later on inside
AbstractSQLClause#setParameters - I figured this was preferable to larger-scale
refactoring.

Fixes #1839