Handle unique null in SimpleExpression in / notIn
# Context

`SimpleExpression.in` and `SimpleExpression.notIn` will transform collection/array with a unique element into `eq(element)` and `ne(element)` respectively.
The problem is that when this unique element is `null`, it will throw an `IllegalArgumentException` with the following message:

```
eq(null) is not allowed. Use isNotNull() instead
```

or

```
ne(null) is not allowed. Use isNotNull() instead
```

The query should not be changed and fail because of this transformation, but instead let the backing datastore run the query.

# What was done

Changed the following methods from `SimpleExpression` to prevent failure from QueryDSL API:
- `in(Collection<? extends T> right)`
- `in(T... right)`
- `notIn(Collection<? extends T> right)`
- `notIn(T... right)`