com.querydsl.sql.StatementOptions - Enable global settings and method chaining
Add com.querydsl.sql.StatementOptions filed into Configuration object so that it can be used when a new Sql Query is created (via the Sql Query's configuration field and the configuration's statementOptions field) in order to configure the default settings for the JDBC statement used by a Sql Query.

Make the AbstractSQLQuery.setStatementOptions returning the query in order to allow method chaining.

Add AbstractSQLQuery.fetchSize method to set the JDBC fetch size (via setStatementOptions) to be used by the underlying JDBC Statement.